Change navigation font back to normal 1em size and increase vertical padding for touch screen usability.
I initially changed font-size to 1.25em because of what I was seeing on a mobile device. However, it turns out I forgot to set the viewport, so the mobile browser was just zooming out. Now that this is fixed, the font-size was too large on mobile devices, so I reset it back.